# openapi.yaml
openapi: 3.0.3
info:
  title: "LoyaltyGen API"
  description: |
    API RESTful para la plataforma de lealtad de clientes LoyaltyGen.
    Permite gestionar clientes, programas de afinidad y cuentas de puntos de manera flexible y potente.
  version: "1.0.0"
servers:
  - url: /api/v1
    description: Servidor Principal
...
security:
  - BearerAuth: []
paths:
...
  /clients/{client_id}/accounts:
    get:
      tags: [Loyalty Accounts]
      summary: Listar Cuentas de un Cliente
      parameters:
        - $ref: '#/components/parameters/ClientId'
      responses:
        '200':
          description: Una lista de las cuentas de lealtad del cliente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoyaltyAccount'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Loyalty Accounts]
      summary: Crear una Cuenta de Lealtad
      parameters:
        - $ref: '#/components/parameters/ClientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Cuenta creada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyAccount'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /clients/{client_id}/accounts/{account_id}/credit:
    post:
      tags: [Loyalty Accounts]
      summary: Acreditar Puntos
      parameters:
        - $ref: '#/components/parameters/ClientId'
        - $ref: '#/components/parameters/AccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditDebitRequest'
      responses:
        '200':
          description: Puntos acreditados. Devuelve la cuenta actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyAccount'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /clients/{client_id}/accounts/{account_id}/debit:
    post:
      tags: [Loyalty Accounts]
      summary: Debitar Puntos
      parameters:
        - $ref: '#/components/parameters/ClientId'
        - $ref: '#/components/parameters/AccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditDebitRequest'
      responses:
        '200':
          description: Puntos debitados. Devuelve la cuenta actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /clients/{client_id}/accounts/{account_id}/transactions:
    get:
      tags: [Loyalty Accounts]
      summary: Listar Transacciones de una Cuenta
      description: |
        Obtiene una lista paginada de transacciones con soporte para filtrado por tipo y rango de fechas.
      parameters:
        - $ref: '#/components/parameters/ClientId'
        - $ref: '#/components/parameters/AccountId'
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: next_cursor
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          description: Fecha de inicio en formato ISO 8601 (inclusive).
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          description: Fecha de fin en formato ISO 8601 (inclusive).
          schema:
            type: string
            format: date-time
        - name: transaction_type
          in: query
          description: Filtrar por tipo de transacción. Si no se proporciona, devuelve todas.
          schema:
            type: string
            enum: [credit, debit]
      responses:
        '200':
          description: Una lista paginada de transacciones.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  paging:
                    $ref: '#/components/schemas/Paging'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /clients/{client_id}/balance:
    get:
      tags: [Loyalty Accounts]
      summary: Obtener Todos los Saldos de un Cliente
      description: Consulta rápida de todos los saldos del cliente leyendo el campo desnormalizado.
      parameters:
        - $ref: '#/components/parameters/ClientId'
      responses:
        '200':
          description: Un mapa con los IDs de cuenta y sus saldos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllBalancesResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /clients/{client_id}/accounts/{account_id}/balance:
    get:
      tags: [Loyalty Accounts]
      summary: Obtener Saldo de una Cuenta Específica
      parameters:
        - $ref: '#/components/parameters/ClientId'
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: El saldo de puntos de la cuenta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # --------------------
  # Audit Trail Endpoints
  # --------------------
...
components:
  schemas:
...
    RelationshipType:
      type: string
      enum: [spouse, child, parent, sibling, friend, other]
      description: |
        Tipo de relación en el círculo de afinidad familiar:
        - spouse: Cónyuge
        - child: Hijo/a
        - parent: Padre/Madre
        - sibling: Hermano/a
        - friend: Amigo/a
        - other: Otra relación
...
    FamilyCircleConfig:
      type: object
      nullable: true
      description: "Configuración de permisos de círculo familiar para una cuenta de lealtad."
      properties:
        allowMemberCredits:
          type: boolean
          description: "Indica si los miembros del círculo pueden generar transacciones de crédito en esta cuenta."
        allowMemberDebits:
          type: boolean
          description: "Indica si los miembros del círculo pueden generar transacciones de débito en esta cuenta."
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          description: "UID del usuario que actualizó la configuración."
    TransactionOriginator:
      type: object
      nullable: true
      description: "Información del cliente que originó la transacción (solo para transacciones de círculo familiar)."
      properties:
        clientId:
          type: string
          description: "ID del cliente que originó la transacción."
        isCircleMember:
          type: boolean
          description: "True si la transacción fue originada por un miembro del círculo, false si fue el titular."
        relationshipType:
          $ref: '#/components/schemas/RelationshipType'
          nullable: true
...
    LoyaltyAccount:
      type: object
      properties:
        id: { type: string }
        account_name: { type: string }
        points: { type: integer }
        familyCircleConfig:
          $ref: '#/components/schemas/FamilyCircleConfig'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Transaction:
      type: object
      properties:
        id: { type: string }
        transaction_type: { type: string, enum: [credit, debit] }
        amount: { type: number, description: "Siempre positivo." }
        description: { type: string }
        originatedBy:
          $ref: '#/components/schemas/TransactionOriginator'
        timestamp: { type: string, format: date-time }
    
    # Audit Trail Entities
...
    CreateAccountRequest:
      type: object
      required: [account_name]
      properties:
        account_name: { type: string }
    CreditDebitRequest:
      type: object
      required: [amount]
      properties:
        amount: { type: number, minimum: 1 }
        description: { type: string }

    # Responses
    AllBalancesResponse:
      type: object
      description: "Un mapa con los IDs de cuenta como clave y sus saldos como valor."
      additionalProperties:
        type: integer
      example:
        account-id-1: 100
        account-id-2: 50
    BalanceResponse:
      type: object
      properties:
        points:
          type: integer
          example: 100
...
    Paging:
      type: object
      properties:
        next_cursor: { type: string, nullable: true, description: "Cursor para la siguiente página." }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
  parameters:
    ClientId:
      name: client_id
      in: path
      required: true
      description: ID del cliente.
      schema: { type: string }
...
    AccountId:
      name: account_id
      in: path
      required: true
      description: ID de la cuenta de lealtad.
      schema: { type: string }
...
  responses:
    NotFound:
      description: El recurso especificado no fue encontrado.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: El token de autenticación falta, es inválido o ha expirado.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: La petición está malformada o es inválida.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "ID Token de Firebase Authentication (JWT)"